extend matching for gravel paths; update README
fixes #1

+++ Makefile
@@ -9,13 +9,16 @@ browse-testmap:
 	open http://localhost:8092/map.html
 
 style.json: $(HOME)/Downloads/gravel_overlay.json
-	mv -v $< $@
+	test -f $< && mv -v $< $@ || true
+
+$(HOME)/Downloads/gravel_overlay.json:
+	@true
 
 gravel-style.json: style.json
-	sed -e 's/http:\/\/localhost:8090/https:\/\/map-tiles.m11n.de\/gravel\/vector\/m11n\/data\/v3.json/' style.json > $@
+	sed -e 's/http:\/\/gravel:8080/https:\/\/map-tiles.m11n.de\/gravel\/vector\/m11n\/data\/v3.json/' $< > $@
 
-.PHONY: install-in-brouter-web
-install-in-brouter-web: gravel-style.json
+.PHONY: install-in-bikerouter
+install-in-bikerouter: gravel-style.json
 	cp -v gravel-style.json $(BROUTER_DEV)/brouter-web/layers/mvt/gravel-overlay-style.json
 	@echo "Gravel Overlay style copied to BRouter Web layers directory"
 	@echo "Change to Brouter Web directory an build application:"
